% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_password.R
\name{keep_it_safe}
\alias{keep_it_safe}
\title{Keep it secret, and keep it safe (with a password)}
\usage{
keep_it_safe(method = c("word", "phrase", "sentence", "random"), ...)
}
\arguments{
\item{method}{string. Method to use for password generation.}

\item{...}{pass arguments to password generator functions.}
}
\value{
Returns a string with the generated password.
}
\description{
Password generator using text from the Lord of the Rings novels.
}
\examples{
# generate a password using a random method
keep_it_safe("random")
}
