% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr}
\alias{lotr}
\title{Lord of the Rings text}
\format{
A character vector of length 41088
}
\source{
{J.R.R. Tolkien. 1954. The Lord of the Rings. HarperCollins e-books. \url{https://archive.org/details/tolkien-j.-the-lord-of-the-rings-harper-collins-ebooks-2010/page/n3/mode/2up}}
}
\usage{
lotr
}
\description{
The full text for the Lord of the Rings novel.
}
\keyword{datasets}
