% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_index}
\alias{lotr_index}
\title{Lord of the Rings index}
\format{
A tibble with 62 rows and 7 variables:
\describe{
\item{chapters}{A string denoting chapter number)}
\item{name_orig}{A string denoting the original name of the chapter as it appears in the text}
\item{name_clean}{A string with a clean version of the chapter name}
\item{book}{A string denoting the Book number}
\item{volume}{A string denoting the name of the volume}
\item{line_start}{an integer denoting the line in the text where a given chapter starts}
\item{line_end}{an integer denoting the line in the text where a given chapter ends}
}
}
\usage{
lotr_index
}
\description{
Metadata for \code{\link{lotr}}. Contains the chapter names, volume names, and line
numbers for the text, and can be used to locate these in the text.
}
\keyword{datasets}
