% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lotr_sentences}
\alias{lotr_sentences}
\title{Extract random sentences}
\usage{
lotr_sentences(n = 4)
}
\arguments{
\item{n}{minimum number of words in the sentence.}
}
\description{
Extract random sentences
}
