% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lotr_words}
\alias{lotr_words}
\title{Function to get table of words}
\usage{
lotr_words(min_size = 4, max_size = NULL)
}
\arguments{
\item{min_size}{integer. Minimum number of letters.}

\item{max_size}{integer. Maximum number of letters.}
}
\description{
Extracts all unique words from the Lord of the Rings text.
}
