% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_password.R
\name{is_it_secret}
\alias{is_it_secret}
\alias{is_it_safe}
\title{Is the password secret? Is the password safe?}
\usage{
is_it_secret(pw)

is_it_safe(pw, verbose = T)
}
\arguments{
\item{pw}{string. Password to test against Pwned Passwords database.}

\item{verbose}{logical. Whether or not to print output of strength estimator.}
}
\value{
If the password is not found in the database, the \code{is_it_safe()} function returns
a phrase 'You can keep your secret...'. If the password is present in the database,
the function returns the phrase 'Password found in database X times', where
X is the number of times the provided password appears in the HIBP database.
\code{is_it_safe()} prints the amount of time the password will take to crack at 10 guesses per second,
online with no throttling;
and, if the password is weak, prints warning and suggestions for improvement (if \code{verbose = TRUE}).
}
\description{
Test the strength of a password (is it safe) and whether the password has been involved in a data breach (is it secret).
}
\details{
The \code{is_it_secret()} function queries the \href{https://haveibeenpwned.com/API/v3#PwnedPasswords}{Pwned Passwords API} for matches.
Have I been pwned (HIBP) uses k-anonymity to make sure searched passwords
stay secret. For more details see \href{https://haveibeenpwned.com/API/v3#SearchingPwnedPasswordsByRange}{haveibeenpwned.com}.
The \code{is_it_safe()} function tests the strength of your password using
\href{https://github.com/dropbox/zxcvbn/blob/master/dist/zxcvbn.js}{zxcvbn.js}.
zxcvbn.js has not been updated in a few years, so the estimates of time to
crack are likely overestimates.
}
\examples{
\dontrun{
pw <- "password123"
is_it_secret(pw) # probably not...
is_it_safe(pw) # definitely not...
}
}
\references{
https://dropbox.tech/security/zxcvbn-realistic-password-strength-estimation
}
