% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_password.R
\name{random_words}
\alias{random_words}
\alias{phrase_method}
\alias{sentence_method}
\title{Password generator}
\usage{
random_words(n = 4, cap = c("original", "none", "title"), sep = "-", ...)

phrase_method(n = 6, sep = "")

sentence_method(n = 6)
}
\arguments{
\item{n}{integer. Number of words.}

\item{cap}{Capitalisation of words. Options are to keep as they appear in the text,
no capitalisation, and capitalise each word.}

\item{sep}{What is used to separate the words. Defaults to space.}

\item{...}{pass arguments to \code{\link{lotr_words}}}
}
\description{
Various methods for generating secure passwords.
}
